requirements.txt
# Core LLM framework
langchain>=0.1.16
langchain-community>=0.0.36

# Vector database (Pinecone v3 – serverless)
pinecone>=3.0.0

# Embeddings
sentence-transformers>=2.2.2
transformers>=4.38.0

# PDF loading
pypdf>=4.0.0

# App / environment
python-dotenv>=1.0.1
flask>=3.0.0

# Utilities
numpy>=1.24.0
tqdm>=4.66.0